Visual Directory Scanner
This script uses GoBuster to enumerate web directories then takes snap shots of the page when it gets 200s (you may want to add 403/302/etc). Last it will create a webpage with all the captured images so you can view the scan results visually in a quick fashion. This also comes in hand if you proxies the go buster scan so you can frequent your results without visiting the site. Spoofing UAs and IPs would also be easy with the structure of the loop.